requirements.txt
mysql-connector-python
pandas
boto3
sqlalchemy
python-dotenv
